refactor: create reusable Modal component
