+++ frontend/src/components/Modal.jsx
@@ -1,0 +1,41 @@
+import React, { useEffect, useRef } from 'react';
+import './Modal.css';
+
+const Modal = ({ isOpen, onClose, children }) => {
+  const modalContentRef = useRef(null);
+
+  useEffect(() => {
+    if (!isOpen) return;
+
+    const handleClickOutside = (event) => {
+      if (modalContentRef.current && !modalContentRef.current.contains(event.target)) {
+        onClose();
+      }
+    };
+
+    const handleKeyDown = (event) => {
+      if (event.key === 'Escape') {
+        onClose();
+      }
+    };
+
+    document.addEventListener('mousedown', handleClickOutside);
+    document.addEventListener('keydown', handleKeyDown);
+    return () => {
+      document.removeEventListener('mousedown', handleClickOutside);
+      document.removeEventListener('keydown', handleKeyDown);
+    };
+  }, [isOpen, onClose]);
+
+  if (!isOpen) return null;
+
+  return (
+    <div className="modal">
+      <div className="modal-content" ref={modalContentRef}>
+        {children}
+      </div>
+    </div>
+  );
+};
+
+export default Modal;
