+++ frontend/src/components/ReportFormModal.jsx
@@ -1,33 +1,12 @@
-import React, {useEffect, useRef, useState} from 'react';
+import React, {useEffect, useState} from 'react';
 import {useLocalStorage} from '../hooks/useLocalStorage';
+import Modal from './Modal';
 
 const ReportFormModal = ({ isOpen, onClose, onGenerateReport, teams = [] }) => {
     const [startDate, setStartDate] = useState('');
     const [endDate, setEndDate] = useState('');
     const [selectedTeams, setSelectedTeams] = useLocalStorage('reportSelectedTeams', []);
-    const modalContentRef = useRef(null);
 
-    useEffect(() => {
-        const handleClickOutside = (event) => {
-            if (modalContentRef.current && !modalContentRef.current.contains(event.target)) {
-                onClose();
-            }
-        };
-
-        const handleKeyDown = (event) => {
-            if (event.key === 'Escape') {
-                onClose();
-            }
-        };
-
-        document.addEventListener('mousedown', handleClickOutside);
-        document.addEventListener('keydown', handleKeyDown);
-
-        return () => {
-            document.removeEventListener('mousedown', handleClickOutside);
-            document.removeEventListener('keydown', handleKeyDown);
-        };
-    }, [onClose]);
 
     useEffect(() => {
         if (isOpen) {
@@ -60,8 +39,7 @@ const ReportFormModal = ({ isOpen, onClose, onGenerateReport, teams = [] }) => {
     if (!isOpen) return null;
 
     return (
-        <div className="modal">
-            <div className="modal-content" ref={modalContentRef}>
+        <Modal isOpen={isOpen} onClose={onClose}>
                 <form onSubmit={handleSubmit}>
                     <label>
                         Start date
@@ -100,8 +78,7 @@ const ReportFormModal = ({ isOpen, onClose, onGenerateReport, teams = [] }) => {
                         <button type="button" onClick={onClose}>Close</button>
                     </div>
                 </form>
-            </div>
-        </div>
+        </Modal>
     );
 };
