+++ frontend/src/components/settings/UserModal.jsx
@@ -1,6 +1,7 @@
-import React, {useEffect, useRef, useState} from 'react';
+import React, {useEffect, useState} from 'react';
 import {useApi} from '../../hooks/useApi';
 import {toast} from "react-toastify";
+import Modal from '../Modal';
 
 const UserModal = ({ isOpen, onClose, editingUser }) => {
     const [newUserData, setNewUserData] = useState({
@@ -10,7 +11,6 @@ const UserModal = ({ isOpen, onClose, editingUser }) => {
         password: '',
         telegram_username: '',
     });
-    const modalContentRef = useRef(null);
     const { apiCall } = useApi();
 
     useEffect(() => {
@@ -25,18 +25,6 @@ const UserModal = ({ isOpen, onClose, editingUser }) => {
         }
     }, [editingUser]);
 
-    useEffect(() => {
-        const handleClickOutside = (event) => {
-            if (modalContentRef.current && !modalContentRef.current.contains(event.target)) {
-                onClose();
-            }
-        };
-
-        document.addEventListener('mousedown', handleClickOutside);
-        return () => {
-            document.removeEventListener('mousedown', handleClickOutside);
-        };
-    }, [onClose]);
 
     const handleUserFormSubmit = async (e) => {
         e.preventDefault();
@@ -55,8 +43,7 @@ const UserModal = ({ isOpen, onClose, editingUser }) => {
     if (!isOpen) return null;
 
     return (
-        <div className="modal">
-            <div className="modal-content" ref={modalContentRef}>
+        <Modal isOpen={isOpen} onClose={onClose}>
                 <form onSubmit={handleUserFormSubmit}>
                     <label>
                         Name
@@ -103,8 +90,7 @@ const UserModal = ({ isOpen, onClose, editingUser }) => {
                         <button type="button" onClick={onClose}>Close</button>
                     </div>
                 </form>
-            </div>
-        </div>
+        </Modal>
     );
 };
